+++ src/main/java/org/pgcodekeeper/core/database/ms/loader/MsJdbcLoader.java
@@ -70,15 +70,20 @@ public final class MsJdbcLoader extends AbstractJdbcLoader<MsDatabase> {
     }
 
     @Override
-    public void preLoad() {
+    public void preLoad() throws IOException, InterruptedException {
+        if (isPreloaded) {
+            return;
+        }
         try (Connection connection = connector.getConnection();
              Statement statement = connection.createStatement()) {
             queryCheckMsVersion(statement);
-        } catch (InterruptedException e) {
+            isPreloaded = true;
+        } catch (InterruptedException ex) {
             Thread.currentThread().interrupt();
-            throw new IllegalArgumentException(e);
+            throw ex;
         } catch (Exception e) {
-            throw new IllegalArgumentException(e);
+            throw new IOException(Messages.Connection_DatabaseJdbcAccessError.formatted(getCurrentLocation(),
+                    e.getLocalizedMessage()), e);
         }
     }
 
@@ -123,6 +128,7 @@ public final class MsJdbcLoader extends AbstractJdbcLoader<MsDatabase> {
 
             LOG.info(Messages.JdbcLoader_log_succes_queried);
         } catch (InterruptedException ex) {
+            Thread.currentThread().interrupt();
             throw ex;
         } catch (Exception e) {
             // connection is closed at this point